requirements.txt
alabaster==1.0.0
babel==2.17.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
docutils==0.21.2
git-filter-repo==2.47.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
gspread-dataframe==4.0.0
idna==3.10
imagesize==1.4.1
iniconfig==2.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
roman-numerals-py==3.1.0
rsa==4.9
six==1.17.0
snowballstemmer==2.2.0
tzdata==2025.1
urllib3==2.3.0
